+++ app_modules/Libri/controller.py
@@ -17,8 +17,15 @@ class LibriController():
         model_data = list(mongo.db['libri'].find().limit(limit))  # Accesso alla collezione 'libri' nel database 'libreria'
         return model_data
     
+    def exists(self, request_id: str):
+        # Verifica se il cliente con l'ID specificato esiste nel database
+        return mongo.db['libri'].count_documents({"_id": ObjectId(request_id)}) > 0
+    
     def get(self, request_id: str):
-        return self.exists(request_id)
+        if self.exists(request_id):
+            return mongo.db['libri'].find_one({"_id": ObjectId(request_id)})
+        else:
+            return None
     
     def create(self, request: Libri):
         if not request._id:
@@ -29,25 +36,20 @@ class LibriController():
 
 
     def update(self, request_id: str, request: Libri):
-        collection_name = self.model.collection_name()
-        document = mongo.db[collection_name].find_one_and_update(
-            {"_id": ObjectId(request_id)},
-            {"$set": request.as_dict()},
-        )
-        return document
-
+        if self.exists(request_id):
+            collection_name = self.model.collection_name()
+            document = mongo.db[collection_name].find_one_and_update(
+                {"_id": ObjectId(request_id)},
+                {"$set": request.as_dict()},
+            )
+            # return document
+        else:
+            return None
  
     def delete(self, request_id: str):
-        collection_name = self.model.collection_name()
-        mongo.db[collection_name].delete_one({"_id": ObjectId(request_id)})
-        return None
-
-
-
-    # def create(self, request: Libri):
-    #     request._id = (
-    #         mongo.db[self.model.collection_name()]
-    #         .insert_one(request.as_dict())
-    #         .inserted_id
-    #     )
-    #     return request
+        if self.exists(request_id):
+            collection_name = self.model.collection_name()
+            mongo.db[collection_name].delete_one({"_id": ObjectId(request_id)})
+            return True
+        else:
+            return False
